Moved the state enum into it's own extension. Cleaned up the redundant function modifiers and comments.
